refactor(workflows): brace 抽出と行解析を _brace.js に 1 本化する
draft PR の conformance と structure が挙げた 3 点を潰す。

extractBracedBody は _brace.js と audit.routing.test.js に同じループが 2 本
あった。引数だけが marker と name で違い、ふるまいは同一だった。行解析の
parseRoutingLikeConst も、meta-contract.test.js が正規表現をキャプチャ 1 つ
違いで書き写していた。両方を _brace.js に集約し、呼び出し側 2 本を import に
切り替えた。FOCUS のキー抽出は Object.keys(parseRoutingLikeConst(source,
"FOCUS")) になる。

.ja/workflows/audit.js の invalid-focus の why は英語のままだった。同ファイル
47 行の no-repo は日本語で、rules/conventions/MIRROR.md は message string を
翻訳対象と定める。日本語へ直した。
